+++ lib/src/serialazy.ts
@@ -9,7 +9,7 @@ import { JsonMap } from './types/json_type';
  * @param serializable Serializable type
  * @returns JSON-compatible type which can be safely passed to `JSON.serialize`
  */
-export function deflate(serializable: any): JsonMap {
+function deflate(serializable: any): JsonMap {
 
     let serialized: JsonMap;
 
@@ -41,7 +41,7 @@ export function deflate(serializable: any): JsonMap {
  * @param serialized JSON-compatible object (e.g. returned from `JSON.parse`)
  * @returns Serializable instance
  */
-export function inflate<T>(ctor: Constructable<T>, serialized: JsonMap): T {
+function inflate<T>(ctor: Constructable<T>, serialized: JsonMap): T {
 
     let classInstance: T;
 
@@ -75,7 +75,7 @@ export function inflate<T>(ctor: Constructable<T>, serialized: JsonMap): T {
  * Check if target is a serializable value or type
  * @param target Target to check
  */
-export function isSerializable(target: any): boolean {
+function isSerializable(target: any): boolean {
 
     if (target === undefined) { // undefined is not a valid JSON value
         return false;
@@ -105,7 +105,7 @@ export function isSerializable(target: any): boolean {
  * Asserts that target is a serializable value or type
  * @param target Target to check
  */
-export function assertSerializable(target: any): void {
+function assertSerializable(target: any): void {
 
     if (!isSerializable(target)) {
         throw new Error(
@@ -116,34 +116,7 @@ export function assertSerializable(target: any): void {
 
 }
 
-/**
- * Traverse recursively all serializable properties of `destination`, merge them
- * with corresponding properties of `source` and return resulting object.
- * NOTE: This function mutates `destination`
- * @param destination Destination serializable class instance
- * @param source Source class instance or plain object (may be non-serializable) to take property values from
- * @returns Destination class instance
- */
-export function deepMerge<T>(destination: T, source: T): T {
-
-    if (destination === null || destination === undefined) {
-        throw new Error('Expecting `destination` to be not null/undefined');
-    }
-
-    assertSerializable(destination);
-    const meta = Metadata.getOwnOrInheritedMetaFor(Object.getPrototypeOf(destination));
-
-    if (source !== null && source !== undefined) {
-        try {
-            meta.aggregateSerializers().forEach(serializer => serializer.assign(destination, source));
-        } catch (error) {
-            throw new Error(`Unable to perform a deep property merge for instance of "${meta.name}": ${error.message}`);
-        }
-    }
-
-    return destination;
-
-}
+export { inflate, deflate, isSerializable };
 
 // Export decorators
 export { default as Serialize} from './decorators/serialize';